fix - fix some bugs

+++ src/1.Domain/AYweb.Domain/Models/Notification/Entities/UserNotification.cs
@@ -6,11 +6,11 @@ public class UserNotification : Entity<long>
 {
     public bool IsSeen { get; private set; }
 
-    public int NotificationId { get; private set; }
+    public long NotificationId { get; private set; }
 
     public Notification Notification { get; private set; }
 
-    public int UserId { get; private set; }
+    public long UserId { get; private set; }
 
     public User.Entities.User User { get; private set; }
 
@@ -18,13 +18,13 @@ public class UserNotification : Entity<long>
 
     #region Constructors And Factories
     private UserNotification() { }
-    public UserNotification(int notificationId, int userId)
+    public UserNotification(long notificationId, long userId)
     {
         NotificationId = notificationId;
         UserId = userId;
         CreateAt = DateTime.Now;
     }
-    public static UserNotification Create(int notificationId, int userId)
+    public static UserNotification Create(long notificationId, long userId)
     {
         return new UserNotification(notificationId, userId);
     }
